+++ server/src/common/config/passport.ts
@@ -4,12 +4,10 @@ const GitHubStrategy = require('passport-github').Strategy;
 
 module.exports = () => {
   passport.serializeUser((user, done) => {
-    // console.log('serializeUser', user);
-    done(null, user.id);
+    done(null, user);
   });
 
   passport.deserializeUser((user, done) => {
-    // console.log('deserializeUser');
     done(null, user);
   });
 
@@ -20,11 +18,7 @@ module.exports = () => {
         clientSecret: 'efc948fa6c7f8f29a31309c774f21c2749d5ea85',
         callbackURL: 'http://127.0.0.1:3000/oauth/github/callback'
       },
-      (accessToken, refreshToken, profile, cb) => {
-        // console.log(accessToken, refreshToken, profile);
-        // User.findOrCreate({ githubId: profile.id }, function (err, user) {
-        //   return cb(err, user);
-        // });
+      async (accessToken, refreshToken, profile, cb) => {
         return cb(null, profile);
       }
     )
