+++ server/src/router/api/oauth.ts
@@ -1,31 +1,13 @@
 import express from 'express';
 import { createConnection } from 'typeorm';
+import OauthController from '@controller/oauth-controller';
 
 const passport = require('passport');
-const jwt = require('jsonwebtoken');
 
 const router = express.Router();
-interface user {
-  username: string;
-}
-declare module 'express' {
-  interface Request {
-    user?: user;
-  }
-}
+
 router.get('/github/login', passport.authenticate('github'));
 
-router.get('/github/callback', passport.authenticate('github', { failureRedirect: '/' }), (req: express.Request, res: express.Response) => {
-  console.log(req.user.username);
-  const token = jwt.sign(
-    {
-      userId: req.user.username
-    },
-    'SECRET',
-    { expiresIn: 60 }
-  );
-  res.cookie('jwt', token);
-  res.redirect('/');
-});
+router.get('/github/callback', passport.authenticate('github', { failureRedirect: '/' }), OauthController.OauthCallback);
 
 export default router;
