+++ server/tsconfig.json
@@ -22,7 +22,7 @@
       "@repository/*": ["src/repository/*"],
       "@router/*": ["src/router/*"],
       "@service/*": ["src/service/*"],
-      "@config/*":["src/config/*"],
+      "@config/*":["src/common/config/*"],
     }
   },
 }
